Case #7318, Case #7319 - Better init and error handling
Case #7318: Don't require that CONDOR_AGENT_SUBMIT_DIR has to be defined if the agent isn't being asked to do proxy submissions to the queue. Moved the check for this directory out of the __init__() method for the HTTP server handler class and in to a sub-routine that gets called at submission time.

Case  #7319: If CONDOR_AGENT_SUBMIT_DIR cannot be created the failure now throws an exception in the submit() method, which is caught and sent back to the caller as a 500 status message. Much better. Means failures to submit aren't ever silent.