Ensure submission cleaner thread doesn't die
On startup the thread checks for CONDOR_AGENT_SUBMIT_DIR -- if this isn't defined the Agent shouldn't let the cleaner thread exit. Instead it should just warn that it's not doing cleanup at this point in time and sleep as it usually does.

This lets the user correct the problem without having to restart the entire Agent subsystem.

To support this I added the concept of a "default" return value for a config value query. If the query returns None the default supplied to the method is returned instead of None. This lets functions have the default be something like a valid, but empty, string instead of None. Makes it easier to code if you can assume it's always a string that might be empty and you don't have to deal with the None case.