Case #7403: Make sure the Agent always returns to its IWD
There was some ancedotal evidence that the agent was losing it's place on the filesystem over the course of many failed submissions attempts from jobs with mal-formed submission tickets. These changes add additional debug information to aid in tracking the problem and some code that forces a return to the initial working directory after servicing a request.