Case 7403: Submissions fail after a period of time
Tracked the problem down to using os.getcwd() -- this call is not thread-safe. It queries a process-level global. This means a thread doing an os.chdir() can change the value that *another* thread gets when it call os.getcwd().

Also added detailed documentation for accessing the REST API for submissions with a nice example that uses curl from the command line to do a "remote" submission via the API.