Case #8236: Race condition in cleanup on slow Condor startup
An overly burdened scheduler can take a while to return jobs when it first starts up, despite actually having jobs in the queue. The cleanup thread needs to sleep when it first starts up so it doesn't query a scheduler in this state, think it's empty, and wipe all the temporary directories of the jobs that are in the scheduler's queue.